+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+* A checkout's spec is now immutable when it has selected a volume to use from a pool.
+
 ### Fixed
 
 * Copy the `AccessModes` requested in a checkout to the checkout's volume.